moved service to it`s own dir
