+++ ms-api/src/test/java/ua/ms/services/MailAlertMessageQueueTest.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ActiveProfiles;
-import ua.ms.service.mq.impl.MailAlertService;
+import ua.ms.service.mq.impl.mail.MailAlertService;
 
 import java.io.Serializable;
